+++ src/protontricks/cli/main.py
@@ -15,7 +15,7 @@ import sys
 from .. import __version__
 from ..flatpak import (FLATPAK_BWRAP_COMPATIBLE_VERSION,
                        get_running_flatpak_version)
-from ..gui import select_steam_app_with_gui, show_text_dialog
+from ..gui import prompt_filesystem_access, select_steam_app_with_gui
 from ..steam import (find_legacy_steam_runtime_path, find_proton_app,
                      find_steam_path, get_steam_apps, get_steam_lib_paths)
 from ..util import run_command
@@ -204,6 +204,12 @@ def main(args=None):
 
     # 4. Find any Steam library folders
     steam_lib_paths = get_steam_lib_paths(steam_path)
+
+    # Check if Protontricks has access to all the required paths
+    prompt_filesystem_access(
+        paths=[steam_path, steam_root] + steam_lib_paths,
+        show_dialog=args.no_term
+    )
 
     # 5. Find any Steam apps
     steam_apps = get_steam_apps(