+++ src/protontricks/flatpak.py
@@ -1,8 +1,11 @@
-from pathlib import Path
-
 import configparser
+import json
 import logging
 import re
+import shlex
+from pathlib import Path
+
+from .config import get_config
 
 __all__ = (
     "FLATPAK_BWRAP_COMPATIBLE_VERSION", "FLATPAK_INFO_PATH",